fix bookmarks stuff

+++ plugin.js
@@ -9,14 +9,7 @@ function registerPlugins(data) {
         //addOptionsToUI(manifest.options, manifest.id);
         if (!path.endsWith('/')) path+='/';
         //console.log(path+manifest.script)
-        var bm = bookmarks.matchAndAccess(path+manifest.script);
-        try {
-            functions.push([require(path+manifest.script), manifest.id]);
-        } catch(e) {
-            bookmarks.release(bm);
-            throw e;
-        }
-        bookmarks.release(bm);
+        functions.push([require(path+manifest.script), manifest.id]);
         // note - server will need to restart to load changes in plugin
         
     }
@@ -66,30 +59,22 @@ function registerPlugins(data) {
 function copyFileSync(source, target) {
     var targetFile = target;
     // If target is a directory, a new file with the same name will be created
-    var bm = bookmarks.matchAndAccess(target);
     if (fs.existsSync(target)) {
         if (fs.lstatSync(target).isDirectory()) {
             targetFile = path.join(target, path.basename(source));
         }
     }
-    bookmarks.release(bm);
-    var bm = bookmarks.matchAndAccess(targetFile);
-    var bm2 = bookmarks.matchAndAccess(source);
+    var bm = bookmarks.matchAndAccess(source);
     fs.writeFileSync(targetFile, fs.readFileSync(source));
     bookmarks.release(bm);
-    bookmarks.release(bm2);
 }
 
 function copyFolderRecursiveSync(source, targetFolder) {
     //console.log('copy', source, targetFolder);
     var files = [];
-    var bm = bookmarks.matchAndAccess(targetFolder);
     if (!fs.existsSync(targetFolder)) fs.mkdirSync(targetFolder);
-    bookmarks.release(bm);
-    var bm = bookmarks.matchAndAccess(targetFolder);
     if (fs.lstatSync(source).isDirectory()) {
         files = fs.readdirSync(source);
-        bookmarks.release(bm);
         for (var i=0; i<files.length; i++) {
             var curSource = path.join(source, files[i]);
             var bm = bookmarks.matchAndAccess(curSource);
@@ -100,28 +85,20 @@ function copyFolderRecursiveSync(source, targetFolder) {
             }
             bookmarks.release(bm);
         }
-        return;
     }
-    bookmarks.release(bm);
 }
 
 function deleteFolder(folder) {
-    var bm = bookmarks.matchAndAccess(folder);
     if (fs.lstatSync(folder).isDirectory()) {
         var files = fs.readdirSync(folder);
-        bookmarks.release(bm);
         for (var i=0; i<files.length; i++) {
             var curSource = path.join(folder, files[i]);
-            var bm = bookmarks.matchAndAccess(curSource);
             if (fs.lstatSync(curSource).isDirectory()) {
                 deleteFolder(curSource);
             }
             fs.unlinkSync(curSource);
-            bookmarks.release(bm);
         }
-        return;
     }
-    bookmarks.release(bm);
 }
 
 function importPlugin(path, config) {
@@ -136,30 +113,35 @@ function importPlugin(path, config) {
             break;
         }
     }
-    var bm = bookmarks.matchAndAccess(global.path.join(eApp.getPath('userData'), "plugins", manifest.id));
     if (global.fs.existsSync(global.path.join(eApp.getPath('userData'), "plugins", manifest.id))) {
         deleteFolder(global.path.join(eApp.getPath('userData'), "plugins", manifest.id));
     }
-    bookmarks.release(bm);
-    var bm = bookmarks.matchAndAccess(global.path.join(eApp.getPath('userData'), "plugins"));
     if (!global.fs.existsSync(global.path.join(eApp.getPath('userData'), "plugins"))) {
         global.fs.mkdirSync(global.path.join(eApp.getPath('userData'), "plugins"));
     }
-    bookmarks.release(bm);
     copyFolderRecursiveSync(path, global.path.join(eApp.getPath('userData'), "plugins", manifest.id));
     if (!inListAlready) config.push({id:manifest.id, config:[]});
     return config;
 }
 
-function removePlugin(id, config) {
-    if (fs.existsSync(global.path.join(eApp.getPath('userData'), "plugins", id))) {
-        deleteFolder(path.join(eApp.getPath('userData'), "plugins", id));
-    }
+function removePlugin(id, serverConfig, config) {
     for (var i=0; i<config.length; i++) {
         if (config[i].id === id) {
             config.splice(i, 1);
             break;
         }
+    }
+    var usedOtherPlaces = false;
+    for (var j=0; j<serverConfig.plugin.length; j++) {
+        for (var i=0; i<serverConfig.plugin[j].length; i++) {
+            if (serverConfig.plugin[j][i].id === id) {
+                usedOtherPlaces = true;
+                break;
+            }
+        }
+    }
+    if (!usedOtherPlaces && fs.existsSync(global.path.join(eApp.getPath('userData'), "plugins", id))) {
+        deleteFolder(path.join(eApp.getPath('userData'), "plugins", id));
     }
     return config;
 }
